requirements.txt
streamlit
tensorflow
cv2
keras
pillow
numpy
pandas
